+++ as/src/base/scan.c
@@ -642,8 +642,8 @@ basic_scan_job_slice(as_job* _job, as_partition_reservation* rsv)
 		as_index_reduce_live(tree, basic_scan_job_reduce_cb, (void*)&slice);
 	}
 	else {
-		uint32_t sample_count = (uint32_t)
-				((as_index_tree_size(tree) * (uint64_t)job->sample_pct) / 100);
+		uint64_t sample_count =
+				((as_index_tree_size(tree) * job->sample_pct) / 100);
 
 		as_index_reduce_partial_live(tree, sample_count,
 				basic_scan_job_reduce_cb, (void*)&slice);
